Struct discovery1::DirectoryListItems

pub struct DirectoryListItems {
    pub kind: Option<String>,
    pub discovery_rest_url: Option<String>,
    pub description: Option<String>,
    pub icons: Option<DirectoryListItemsIcons>,
    pub labels: Vec<String>,
    pub preferred: Option<bool>,
    pub discovery_link: Option<String>,
    pub version: Option<String>,
    pub title: Option<String>,
    pub documentation_link: Option<String>,
    pub id: Option<String>,
    pub name: Option<String>,
}

The individual directory entries. One entry per api/version pair.

This type is not used in any activity, and only used as part of another schema.

Fields

kind

The kind for this response.

discovery_rest_url

The URL for the discovery REST document.

description

The description of this API.

icons

Links to 16x16 and 32x32 icons representing the API.

labels

Labels for the status of this API, such as labs or deprecated.

preferred

True if this version is the preferred version to use.

discovery_link

A link to the discovery document.

version

The version of the API.

title

The title of this API.

documentation_link

A link to human readable documentation for the API.

id

The id of this API.

name

The name of the API.